chore: clear temp files directly in operations instead of injecting a hook
